Improved the corresponding function of Item and changed Cart.

+++ src/main/java/com/csu/petstorepro/petstore/entity/Cart.java
@@ -33,7 +33,7 @@ public class Cart implements Serializable {
 
     private Integer quantity;
 
-    private BigDecimal listprice;
+    private Double listprice;
 
-    private BigDecimal total;
+    private Double total;
 }
